% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_uuid.R
\name{impute_uuid}
\alias{impute_uuid}
\title{Impute UUID for missing values}
\usage{
impute_uuid(x, prefix = NULL)
}
\arguments{
\item{x}{a character vector.}
}
\description{
This function imputes a v4 UUID by refernce into a character
vector. Be careful. This operation cannot be undone!
}
