% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{new_v4}
\alias{new_v4}
\alias{new_v7}
\title{Generate new UUIDs}
\usage{
new_v4(n)

new_v7(n)
}
\arguments{
\item{n}{the number of uuids to generate}
}
\description{
Note that if creating a v7 UUID, \code{uuid::UUIDgenerate(use.time = TRUE, output = "string")} is faster whereas a v4 UUID is ~20x faster using this implementation.
}
