% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_uuid.R
\name{short_uuid}
\alias{short_uuid}
\title{Generate new Short UUID}
\usage{
short_uuid(uuid, short = c("flickr_base58", "bitcoin58"))
}
\arguments{
\item{n}{the number of short uuids to generate}
}
\description{
Generate new Short UUID
}
