% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uuid_to_short.R
\name{uuid_to_short}
\alias{uuid_to_short}
\title{Convert uuid to short}
\usage{
uuid_to_short(uuid, short = c("flickr_base58", "bitcoin58"))
}
\arguments{
\item{uuid}{vector of uuids}
}
\description{
Convert uuid to short
}
