% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_uuid.R
\name{new_short}
\alias{new_short}
\title{Generate new Short UUID}
\usage{
new_short(n, short = c("flickr_base58", "bitcoin58"))
}
\arguments{
\item{n}{the number of short uuids to generate}

\item{short}{what alphabet to use}
}
\description{
Generate new Short UUID
}
