% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_uuid.R
\name{short_to_uuid}
\alias{short_to_uuid}
\title{Convert short to uuid}
\usage{
short_to_uuid(short_uuid, short = c("flickr_base58", "bitcoin58"))
}
\arguments{
\item{short_uuid}{vector of short-uuids}

\item{short}{what alphabet to use}
}
\description{
Convert short to uuid
}
\seealso{
\code{\link[=new_short]{new_short()}}
}
