% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{short_uuid}
\alias{short_uuid}
\alias{short_flickr_base58}
\alias{short_bitcoin58}
\title{Generate new Short UUID}
\usage{
short_flickr_base58(n)

short_bitcoin58(n)
}
\arguments{
\item{n}{the number of short uuids to generate}
}
\description{
Generate new Short UUID
}
